% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add}
\alias{add}
\title{add two things}
\usage{
add(x, y)
}
\arguments{
\item{x}{something that can be added to y}

\item{y}{something that can be added to x}
}
\value{
x + y
}
\description{
add two things
}
