% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{apply_named_args}
\alias{apply_named_args}
\title{apply with named arguments}
\usage{
apply_named_args(X, MARGIN, FUN)
}
\arguments{
\item{X}{see apply}

\item{MARGIN}{see apply}

\item{FUN}{see apply}
}
\value{
see apply
}
\description{
apply with named arguments
}
