% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_for_tests.R
\name{calc_approx_final_size_distribution}
\alias{calc_approx_final_size_distribution}
\title{Britton2010 10.1016/j.mbs.2010.01.006}
\usage{
calc_approx_final_size_distribution(R_0, pop_size)
}
\description{
Britton2010 10.1016/j.mbs.2010.01.006
}
