% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summaries_funcs.R
\name{calc_country_attack}
\alias{calc_country_attack}
\title{calculate the attack rate by country}
\usage{
calc_country_attack(res, X, labels)
}
\arguments{
\item{res}{output of main_simulation}

\item{X}{the vector of population sizes in each location/age/risk group}

\item{labels}{data frame outputted by setup_inputs containing the location/
age/risk group for each row in res}
}
\value{
a numeric vector of length n_countries: the attack rate for each country
}
\description{
calculate the attack rate by country
}
