% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summaries_funcs.R
\name{calc_country_time_series}
\alias{calc_country_time_series}
\title{calculate the number of individuals summed over compartments over time for each country}
\usage{
calc_country_time_series(res, X, labels, compartments)
}
\arguments{
\item{res}{output of main_simulation}

\item{X}{the vector of population sizes in each location/age/risk group}

\item{labels}{data frame outputted by setup_inputs containing the location/
age/risk group for each row in res}

\item{compartments}{vector of compartment names to sum over}
}
\value{
a list containing a matrix with n_countries rows. 
Each row contains the number of individuals summed over compartments over time for that country.
}
\description{
calculate the number of individuals summed over compartments over time for each country
}
