% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summaries_funcs.R
\name{calc_cumulative_deaths_time_series}
\alias{calc_cumulative_deaths_time_series}
\title{calculate the cumulative number of deaths over time for each country}
\usage{
calc_cumulative_deaths_time_series(res, X, labels)
}
\arguments{
\item{res}{output of main_simulation}

\item{X}{the vector of population sizes in each location/age/risk group}

\item{labels}{data frame outputted by setup_inputs containing the location/
age/risk group for each row in res}
}
\value{
a list containing a matrix with n_countries rows. 
Each row contains the incidence over time for that country.
}
\description{
calculate the cumulative number of deaths over time for each country
}
