% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summaries_funcs.R
\name{calc_incidence_vaccinated_peak_times_attack_rates}
\alias{calc_incidence_vaccinated_peak_times_attack_rates}
\title{calculate the incidence, number of vaccinated individuals, peak times
and attack rates by country}
\usage{
calc_incidence_vaccinated_peak_times_attack_rates(res, X, labels)
}
\arguments{
\item{res}{output of main_simulation}

\item{X}{the vector of population sizes in each location/age/risk group}

\item{labels}{data frame outputted by setup_inputs containing the location/
age/risk group for each row in res}
}
\value{
a list with four entries.
incidence: the incidence over time for each country
vaccinated: the number of vaccinated individuals over time for each country
peakTimes: numeric vector of length n_countries: the peak time for each country
country_attack_rate: numeric vector of length n_countries: the attack rate for each country
}
\description{
calculate the incidence, number of vaccinated individuals, peak times
and attack rates by country
}
