% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summaries_funcs.R
\name{calc_peak_times_and_attack_rates}
\alias{calc_peak_times_and_attack_rates}
\title{calculate the peak time and attack rate by country}
\usage{
calc_peak_times_and_attack_rates(res, X, labels)
}
\arguments{
\item{res}{output of main_simulation}

\item{X}{the vector of population sizes in each location/age/risk group}

\item{labels}{data frame outputted by setup_inputs containing the location/
age/risk group for each row in res}
}
\value{
a list with two entries.
peakTimes: numeric vector of length n_countries: the peak time for each country
country_attack_rate: numeric vector of length n_countries: the attack rate for each country
}
\description{
calculate the peak time and attack rate by country
}
