% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{calibrating_amp_and_travel}
\alias{calibrating_amp_and_travel}
\title{Cluster submission for amp and epsilon calibration}
\usage{
calibrating_amp_and_travel(runName, amp, epsilon, n_runs, tmax, tdiv,
  seasonality_resolution, life_history_params, travel_params, simulation_flags,
  vax_params, vax_production_params, vax_allocation_params, vax_alloc_period,
  seedCountries, seedSizes, seedAges, seedRiskGroups, tdelay, regionDat,
  latitudeDat, requested_stats)
}
\arguments{
\item{runName}{the base of the filename to save outputs to. Will be appended to "outputs/calibration_". DO NOT SPECIFY A FILE EXTENSION - THIS IS HANDLED IN THE FUNCTION}

\item{amp}{the amplitude of seasonal forcing}

\item{epsilon}{connectivity of off-diagonal elements in the travel matrix}

\item{n_runs}{number of simulations to run for this set of parameters}

\item{tmax}{the number of days for which to run the simulation}

\item{tdiv}{the number of timesteps per day}

\item{seasonality_resolution}{averages the seasonality function into buckets of this many time steps. For example, if you'd like 12 blocks of seasonality in this time frame, you should set this parameter to (tmax*tdiv)/12}

\item{life_history_params}{named vector (or list) with the numeric elements
"R0", "TR" (time to recovery) and "LP" (latent period)}

\item{travel_params}{list of parameters relating to travel. This should have the following elements: 1) epsilon, which scales the off-diagonals of the travel matrix}

\item{simulation_flags}{named vector (or list) with the logical elements "normaliseTravel"
and "seasonal"}

\item{vax_params}{named vector (or list) with the numeric elements "efficacy"
and "propn_vax0" (initial proportion of vaccinated individuals; assumed constant
across location, age and risk groups)}

\item{vax_production_params}{temp description to suppress roxygen warnings}

\item{vax_allocation_params}{temp description to suppress roxygen warnings}

\item{seedCountries}{temp description to suppress roxygen warnings}

\item{seedSizes}{temp description to suppress roxygen warnings}

\item{seedAges}{temp description to suppress roxygen warnings}

\item{seedRiskGroups}{temp description to suppress roxygen warnings}

\item{tdelay}{temp description to suppress roxygen warnings}

\item{regionDat}{temp description to suppress roxygen warnings}

\item{latitudeDat}{temp description to suppress roxygen warnings}

\item{requested_stats}{temp description to suppress roxygen warnings}

\item{vax_allocation_period}{temp description to suppress roxygen warnings}
}
\value{
returns TRUE if the routine runs correctly. Will create a .csv file of peak time summaries for this run, and a .png plotting the distribution of peak times by country. The output filenames are:
1) outputs/calibration_***runName***_data.csv
2) outputs/calibration_***runName***_plot.png
}
\description{
Function used to submit jobs to the cluster, solving the global model with specified parameters. This particular function is intended to test different values of amp and epsilon, which is why they are amongst the first arguments. To write your own function which can be submitted to the cluster, simply place the parameters that you want to change as the first few arguments.
}
