% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cum_vax_pool_func_closure}
\alias{cum_vax_pool_func_closure}
\title{closure to make vaccine production function}
\usage{
cum_vax_pool_func_closure(user_specified_cum_vax_pool_func,
  vax_production_params)
}
\arguments{
\item{user_specified_cum_vax_pool_func:}{function of vax_production_params and t
which returns the number of vaccines which have ever existed at time t}

\item{vax_production_params:}{list of parameters for vaccine production}
}
\value{
function of t which returns
the number of vaccines which have ever existed at time t
}
\description{
a closure to make a function which returns a scalar:
}

