% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{density_country_attack_rates}
\alias{density_country_attack_rates}
\title{plot the density strips of the attack rate in each country across simulations}
\usage{
density_country_attack_rates(dat)
}
\arguments{
\item{dat}{a data frame outputted neaten_raw_output_by_country with
"melted=TRUE"}
}
\value{
a ggplot object
}
\description{
plot the density strips of the attack rate in each country across simulations
}
