% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gen_int_sum_int}
\alias{gen_int_sum_int}
\title{generate n_int random integers that sum to sum_int}
\usage{
gen_int_sum_int(n_int, sum_int)
}
\arguments{
\item{n_int}{number of random integers to generate}

\item{sum_int}{integer to which those integers should sum to}
}
\value{
numeric vector of length n_int
}
\description{
generate n_int random integers that sum to sum_int
}

