% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{generate_age_matrix}
\alias{generate_age_matrix}
\title{Age matrix proportions}
\usage{
generate_age_matrix(ages)
}
\arguments{
\item{ages}{the vector of ages}
}
\value{
the same vector but as an nx1 matrix
}
\description{
Given a vector of proportions in each age group, converts this to a matrix
}
