% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{generate_age_matrix_uniform}
\alias{generate_age_matrix_uniform}
\title{Uniform location age matrix}
\usage{
generate_age_matrix_uniform(ages, maxAge)
}
\arguments{
\item{ages}{the vector of age boundaries}

\item{maxAge}{the maximum age}
}
\value{
the matrix of proportions in each age goup
}
\description{
Given a vector of age boundaries and a maximum age, returns a matrix with a uniform age distribution
}

