% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{generate_age_mixing}
\alias{generate_age_mixing}
\title{Age mixing manipulatoin}
\usage{
generate_age_mixing(contactVector, n_ages, ON = TRUE, TRANSPOSE = TRUE)
}
\arguments{
\item{contactVector}{vector of length n_ages*n_ages giving age specific contact rates/durations}

\item{n_ages}{the number of age groups}

\item{ON}{bool, if FALSE, turns off age mixing}

\item{TRANSPOSE}{bool, if TRUE, returns the transpose of the created matrix}
}
\value{
the matrix of age-specific contact rates/durations
}
\description{
Converts a vector of age-mixing rates or contact durations into a matrix. If a matrix is passed, just returns the original matrix. Can be switched off to return a matrix of 1s
}
