% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{generate_contact_matrix}
\alias{generate_contact_matrix}
\title{Age mixing matrix generation}
\usage{
generate_contact_matrix(contactRates, contactDur, n_ages, ON = TRUE,
  TRANSPOSE = TRUE)
}
\arguments{
\item{contactRates}{the vector or matrix of age-specific contact rates}

\item{contactDur}{the vector or matrix of age-specific contact durations}

\item{n_ages}{the number of age groups}

\item{ON}{bool, if FALSE, turns off age mixing}

\item{TRANSPOSE}{bool, if TRUE, returns the transpose of the created matrix}
}
\value{
the matrix of age-specific contact rates
}
\description{
Generates the age-mixing contact matrix taking into account contact rates and durations
}

