% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_for_tests.R
\name{get_final_sizes_with_R_eff}
\alias{get_final_sizes_with_R_eff}
\title{function to simulate the final size distribution for a single country}
\usage{
get_final_sizes_with_R_eff(R_eff, change_efficacy)
}
\arguments{
\item{R_eff}{numeric vector of length 1: effective reproduction number.
R_0 = 1.4}

\item{change_efficacy}{logical: if TRUE, achieve R_eff by vaccinating all
individuals but having vaccine efficacy less than unity.  If FALSE, achieve
R_eff by vaccinating a proportion of individual, with vaccine efficacy equal
to unity.}
}
\value{
number vector of length 100: final sizes of 100 simulations.
}
\description{
function to simulate thefinal size distribution for a single country, for a 
give effective reproduction number.
}
