% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_func_names}
\alias{get_func_names}
\title{return names of all functions in a .R file}
\usage{
get_func_names(filename)
}
\arguments{
\item{filename}{location of .R file}
}
\value{
character vector where each entry is the name of a function in the file
}
\description{
return names of all functions in a .R file
}

