% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_user_options}
\alias{get_user_options}
\title{specify user options for cluster}
\usage{
get_user_options(user)
}
\arguments{
\item{user}{a user identifier so that each user can specify their cluster options.}
}
\value{
a list with the working directory on the network drive; the directory
in which the package code sits; and options for didehpc
}
\description{
specify user options for cluster
}
