% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_vars_from_list_with_check}
\alias{get_vars_from_list_with_check}
\title{extract variables from list, throwing an error if they are not found}
\usage{
get_vars_from_list_with_check(x, var_names)
}
\arguments{
\item{x}{list of variables}

\item{var_names}{character vector containing names of variables to extract}
}
\value{
list of selected variables
}
\description{
extract variables from list, throwing an error if they are not found
}

