% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpFuncs.R
\name{initial_test}
\alias{initial_test}
\title{Initial test function}
\usage{
initial_test(x)
}
\arguments{
\item{x}{the parameter to take}
}
\value{
a useless numeric value
}
\description{
Takes one parameter and does something completely useless
}

