% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initial_test_cpp}
\alias{initial_test_cpp}
\title{Cpp useless function}
\usage{
initial_test_cpp(x)
}
\arguments{
\item{x}{completely meaningless}
}
\value{
something else
}
\description{
Another useless function to test package development
}

