% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_integer_like}
\alias{is_integer_like}
\title{check if a number is integer-type or close enough to an integer}
\usage{
is_integer_like(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{numeric: the number}

\item{tol}{numeric: tolerance}
}
\value{
TRUE if x is integer-type or within tol of an integer, FALSE otherwise
}
\description{
check if a number is integer-type or close enough to an integer
}
