% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{main_simulation}
\alias{main_simulation}
\title{main simulation loop}
\usage{
main_simulation(tmax, tdiv, vax_alloc_period, LD, S0, E0, I0, R0, SV0, EV0, IV0,
  RV0, params, cum_vax_pool_func, vax_allocation_func)
}
\arguments{
\item{tmax}{number of days for which to run simulation}

\item{tdiv}{number of timesteps per day}

\item{vax_alloc_period}{allocate vaccines every this number of timesteps}

\item{LD}{force of infection matrix: multiply by the I vector to get force of infection.
a square matrix of side length maxIndex  = n_countries * n_ages * n_riskgroups}

\item{S0}{numeric vector of length max_Index: initial number of unvaccinated
susceptibles in each location, age, risk group}

\item{E0}{numeric vector of length max_Index: initial number of unvaccinated
exposed}

\item{I0}{numeric vector of length max_Index: initial number of unvaccinated
infectious}

\item{R0}{numeric vector of length max_Index: initial number of unvaccinated
removed}

\item{SV0}{numeric vector of length max_Index: initial number of vaccinated
susceptibles}

\item{EV0}{numeric vector of length max_Index: initial number of vaccinated
exposed}

\item{IV0}{numeric vector of length max_Index: initial number of vaccinated
infectious}

\item{RV0}{numeric vector of length max_Index: initial number of vaccinated
removed}

\item{params}{list of model parameters}

\item{cum_vax_pool_func}{a function of time which gives the numebr of vaccines
ever produced at that time}

\item{vax_allocation_func}{a function of the current state of the simulation which
gives the number of vaccines to allocate to each location, age, risk group}
}
\value{
list containing the elements:
beta: infectivity parameter
S: matrix containing the number of unvaccinated susceptibles in each
location, age, risk group at each timestep.
nrow(S) = n_groups = n_countres * n_ages * n_riskgroups
ncol(S) = tend = tmax * tdiv
E: same as Smat but for unvaccinated exposed
I: same as Smat but for unvaccinated infectious
R: same as Smat but for unvaccinated removed
SV: same as Smat but for vaccinated susceptible
EV: same as Smat but for vaccinated exposed
IV: same as Smat but for vaccinated infectious
RV: same as Smat but for vaccinated removed
vax_pool: numeric vector containing number of vaccines available at each timestep.
}
\description{
main simulation loop
}

