% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_tools.R
\name{make_arg_list}
\alias{make_arg_list}
\title{make argument list to run simulations on or off the cluster}
\usage{
make_arg_list(runs = NULL, submit_fn, obj = NULL)
}
\arguments{
\item{runs}{data frame where each row is a set of parameters for which to run the simulation.
# if there are no variable parameters, set to NULL.}

\item{submit_fn}{character string.  name of function to run on cluster}

\item{obj}{if running on cluster, the output of setup_cluster.  if not running
on cluster, set to NULL.}
}
\value{
an argument list to run simulations on or off the cluster
}
\description{
make argument list to run simulations on or off the cluster
}

