% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_y_label}
\alias{make_y_label}
\title{make a string to label the y-axis}
\usage{
make_y_label(y_label_base, thin_every, thin_integer)
}
\arguments{
\item{y_label_base}{string}

\item{thin_every}{an integer.  If 0, ignore.  Otherwise, the matrix was thinned
every this many columns}

\item{thin_integer}{a logical. If TRUE, the matrix was thinned to integer values of 
times.  If FALSE, ignore}
}
\value{
a string
}
\description{
make a string to label the y-axis
}
