% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Map_vapply}
\alias{Map_vapply}
\title{a version of Map() where we specify the format of the output}
\usage{
Map_vapply(FUN, FUN.VALUE, ...)
}
\arguments{
\item{FUN}{function}

\item{FUN.VALUE}{prespecified type of return value}
}
\value{
vector with prespecified type of return value
}
\description{
a version of Map() where we specify the format of the output
}
