% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_mult}
\alias{matrix_mult}
\title{Matrix multiplication}
\usage{
matrix_mult(x, y)
}
\arguments{
\item{x}{the first matrix}

\item{y}{the second matrix}
}
\value{
the multiplied matrix
}
\description{
Uses RcppEigen to multiply two matrices quickly
}
