% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{model_plots}
\alias{model_plots}
\title{Model visualisations - MAIN FUNCTION}
\usage{
model_plots(res, n_ages, n_riskgroups,
  countryFile = "~/Documents/vaxedemic/data/unified/countries_intersect.csv",
  latitudeFile = "~/Documents/vaxedemic/data/JH_cleaning/CLEAN/latitudes_clean.csv",
  regionFile = "~/Documents/vaxedemic/data/JH_cleaning/CLEAN/regions_clean.csv",
  vitalFile = "~/Documents/vaxedemic/data/unified/demographic_data_intersect.csv")
}
\arguments{
\item{n_ages}{the number of age groups}

\item{n_riskgroups}{the number of risk groups}

\item{countryFile}{the full file path to the clean country name file}

\item{latitudeFile}{the full file path to the clean country latitude file}

\item{regionFile}{the full file path to the clean country by region file}

\item{vitalFile}{the full file path to the clean country demography (ie. needs "N") file}

\item{the}{list of simulation outputs as in \code{\link{run_simulation}}}
}
\description{
TO DO: 1) need to make sure that the denominator is correct and that we are grouping correctly; 
2) need to generalise grouping indices
3) Make each country occupy same space in y-axis in heatmap plot
This is very much a work in progress. Takes the output of \code{\link{run_simulation}} and returns 
a set of incidence curve and heatmaps for per capita incidence by region. Requires data table and ggplot2.
}
