% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{neaten_raw_output_by_country}
\alias{neaten_raw_output_by_country}
\title{format the summary statistic by country over all n_runs}
\usage{
neaten_raw_output_by_country(res, labels, regionDat, latitudeDat)
}
\arguments{
\item{res}{summary statistics outputted by run_simulation.  Should be a list
of length n_runs, where each element is a numeric vector of length n_countries.}

\item{labels}{data frame outputted by setup_inputs containing country names}

\item{regionDat}{data frame containing a column for Location and a column for
the region that location is in. read from data/regions_clean.csv}

\item{latitudeDat}{data frame containing a column for Location and a column for
the latitude of the location. read from data/latitudes_intersect.csv}
}
\value{
a data frame with the summary statistic for each run for
each country; the region and hemisphere each country belongs to; and the latitude
of each country
}
\description{
format the summary statistic by country over all n_runs
}

