% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalise}
\alias{normalise}
\title{normalises a vector so that it sums to 1}
\usage{
normalise(vec)
}
\arguments{
\item{vec}{numeric vector to be normalised}
}
\value{
normalised vector of same length as vec
}
\description{
normalises a vector so that it sums to 1
}
