% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{num2str}
\alias{num2str}
\title{turn number(s) into string for filename, replacing decimal points with "point"}
\usage{
num2str(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
character vector of same length as x
}
\description{
turn number(s) into string for filename, replacing decimal points with "point"
}
