% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_country_incidence}
\alias{plot_country_incidence}
\title{plot incidence time series for a country}
\usage{
plot_country_incidence(output, country_name, thin_every = 7,
  thin_integer = FALSE)
}
\arguments{
\item{output}{list of n matrices, where n is the number of runs.
Each row in each matrix corresponds to a country and each column to a time.
the rownames of the matrices are the country names.}

\item{country_name}{name of country for which to plot time series}

\item{thin_every}{an integer.  If 0, ignore.  Otherwise, thin each matrix
every this many columns}

\item{thin_integer}{a logical. If TRUE, thin each matrix to integer values of 
times.  If FALSE, ignore}
}
\value{
ggplot object with 95% CI and ten runs
}
\description{
plot 95% CI and ten runs
}
