% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_country_time_series}
\alias{plot_country_time_series}
\title{plot time series for a country}
\usage{
plot_country_time_series(output, country_name, y_label, thin_every = 0,
  thin_integer = FALSE, thin_by_sum = FALSE)
}
\arguments{
\item{output}{list of n matrices, where n is the number of runs.
Each row in each matrix corresponds to a country and each column to a time.
the rownames of the matrices are the country names.}

\item{country_name}{name of country for which to plot time series}

\item{y_label}{string. y-axis label}

\item{thin_every}{an integer.  If 0, ignore.  Otherwise, thin each matrix
every this many columns}

\item{thin_integer}{a logical. If TRUE, thin each matrix to integer values of 
times.  If FALSE, ignore}

\item{thin_by_sum}{logical.  IF TRUE, thin by adding together columns (e.g.
if thin_every = 2, add togeher columns 1-2, columns 3-4 etc.)  This is useful
for thinning quantities such as incidence.  If FALSE, ignore.}
}
\value{
ggplot object with 95% CI and ten runs
}
\description{
plot 95% CI and ten runs
}
