% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_peak_times}
\alias{plot_peak_times}
\title{plot the median and 95% quantiles of the peak time in each country across simulations}
\usage{
plot_peak_times(dat)
}
\arguments{
\item{dat}{a data frame outputted by calc_median_ci_by_country.
contains the mean, median, 2.5% and 97.5% percentiles for the peak time in
each country; the region and hemisphere each country belongs to; and the latitude
of each country}
}
\value{
a ggplot object
}
\description{
plot the median and 95% quantiles of the peak time in each country across simulations
}
