% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_to_run_on_cluster.R
\name{postprocessing_country_attack}
\alias{postprocessing_country_attack}
\title{postprocess simulations to plot the country attack rate and save outputs}
\usage{
postprocessing_country_attack(res_list, runName, other_info)
}
\arguments{
\item{res_list}{summary statistics for each run and processed inputs produced by run_fixed_params}

\item{runName}{character string to make filename out of}

\item{other_info}{list which provides any other information needed, such as to calculate the summaries}
}
\description{
postprocess simulations to plot the country attack rate and save outputs
}
