% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_funcs.R
\name{postprocessing_incidence_peak_times_attack_rates_deaths}
\alias{postprocessing_incidence_peak_times_attack_rates_deaths}
\title{postprocess simulations to plot the peak time and attack rates and deaths,
and save outputs for the incidence, peak time and attack rates and deaths}
\usage{
postprocessing_incidence_peak_times_attack_rates_deaths(res_list, runName,
  other_info, output_prefix)
}
\arguments{
\item{res_list}{summary statistics for each run and processed inputs produced by run_fixed_params}

\item{runName}{character string to make filename out of}

\item{other_info}{list which provides any other information needed, such as to calculate the summaries}

\item{output_prefix}{character vector of length 1.  Prefix for output filenames}
}
\description{
postprocess simulations to plot the peak time and attack rates and deaths,
and save outputs for the incidence, peak time and attack rates and deaths
}
