% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_funcs.R
\name{postprocessing_simple_save}
\alias{postprocessing_simple_save}
\title{just save summary statistics from simulations as an .rds file}
\usage{
postprocessing_simple_save(res_list, runName)
}
\arguments{
\item{res_list}{summary statistics for each run and processed inputs produced by run_fixed_params}

\item{runName}{character string to make filename out of}

\item{output_prefix}{character vector of length 1.  Prefix for output filenames}
}
\description{
just save summary statistics from simulations as an .rds file
}
