% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{read_contact_data}
\alias{read_contact_data}
\title{function to construct contact matrix from data}
\usage{
read_contact_data(contact_filename)
}
\arguments{
\item{contact_filename}{character vector of length 1: file where contact
data is located}
}
\value{
list of length n_countries, of square contact matrices of side length
n_ages * n_riskgroups.
for each matrix, contactMatrix[i,j] denotes the amount of influence 
an individual of type i has on an individual of type j, 
where type includes age and risk groups.
}
\description{
function to construct contact matrix from data
}
