% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{read_coverage_data}
\alias{read_coverage_data}
\title{Read and process seasonal vaccine coverage data}
\usage{
read_coverage_data(coverage_filename, labels)
}
\arguments{
\item{coverage_filename}{the vector or matrix of age-specific contact rates}

\item{labels}{a data frame containing the number of individuals in each location, age, risk group}
}
\value{
the proportion of doses distributed to each country
}
\description{
Calculates the proportion of seasonal vaccine doses distributed to each country in 2013
}
