% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{read_latitude_data}
\alias{read_latitude_data}
\title{function to construct latitude matrix from data}
\usage{
read_latitude_data(latitude_filename)
}
\arguments{
\item{latitude_filename}{character vector of length 1: file where latitude
data is located}
}
\value{
matrix with 1 column and nrow = n_countries.
The latitude of each country in degrees.
}
\description{
function to construct latitude matrix from data
}
