% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summaries_funcs.R
\name{return_all_res}
\alias{return_all_res}
\title{return the full results of main_simulation}
\usage{
return_all_res(res)
}
\arguments{
\item{res}{output of main_simulation}
}
\value{
output of main_simulation
}
\description{
return the full results of main_simulation
}

