% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{rounds a numeric vector probabilistically such that its sum (an integer) is preserved}
\usage{
round_preserve_sum(vec)
}
\arguments{
\item{vec}{numeric vector which sums to an integer}
}
\value{
numeric vector of the same length as vec: rounded vector
}
\description{
rounds a numeric vector probabilistically such that its sum (an integer) is preserved
}
