% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{run_and_get_deaths_recovered}
\alias{run_and_get_deaths_recovered}
\title{run the simulation and get a vector of the number of deaths and recovereds
in each country/age/risk group}
\usage{
run_and_get_deaths_recovered(simulation_flags, life_history_params, vax_params,
  sim_params, case_fatality_ratio_vec, X, labels, C3, K, latitudes,
  cum_vax_pool_func, vax_allocation_func, tmax, tdiv, vax_alloc_period)
}
\arguments{
\item{simulation_flags}{named vector (or list) with the logical elements "normaliseTravel"
and "seasonal"}

\item{life_history_params}{named vector (or list) with the numeric elements
"R0", "TR" (time to recovery) and "LP" (latent period)}

\item{vax_params}{named vector (or list) with the numeric elements "efficacy"
and "propn_vax0" (initial proportion of vaccinated individuals; assumed constant
across location, age and risk groups)}

\item{sim_params}{list with the numeric elements n_countries, n_ages,
n_riskgroups and seed_vec.  The last of these is a vector specifying the 
initial number of exposed individuals in each location, age and risk group}

\item{case_fatality_ratio_vec}{vector of case fatality ratio in each location, age, risk group:
length n_countries * n_ages * n_riskgroups}

\item{X}{vector of population size in each location, age, risk group:
length n_countries * n_ages * n_riskgroups}

\item{latitudes}{a matrix with 1 column and n_countries rows, giving the latitude
of each country in degrees.}

\item{cum_vax_pool_func}{a function of time which gives the numebr of vaccines
ever produced at that time}

\item{vax_allocation_func}{a function of the current state of the simulation which
gives the number of vaccines to allocate to each location, age, risk group}

\item{tmax}{the number of days for which to run the simulation}

\item{tdiv}{the number of timesteps per day}

\item{vax_alloc_period}{allocate vaccines once every this many timesteps}

\item{contactMatrix}{either a square matrix with side length n_ages * n_riskgroups,
or a list of such matrices, of length n_countries.  
If a single matrix, the contact matrix is assumed to be constant across countries.
contactMatrix[i,j] denotes the amount of influence an individual of type i
has on an individual of type j, where type includes age and risk groups.}

\item{travelMatrix}{a square matrix with side length n_countries.
travelMatrix[i,j] denotes the proportion of time an individual in country i
spends in country j.  If simulation_flags[["normaliseTravel"]] == TRUE,
the code will row normalise travelMatrix for you.}
}
\value{
list containing the number of deaths and recovereds
in each country/age/risk group
}
\description{
for code testing purposes.
}
