% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{function to set up simulation from inputs, then run}
\usage{
run_simulation(simulation_flags, life_history_params, vax_params,
  seasonality_params, time_params, vax_alloc_period, processed_inputs,
  n_runs = 1, calculate_summaries_func = "return_all_res", other_info)
}
\arguments{
\item{simulation_flags}{named vector (or list) with the logical elements "normaliseTravel"
and "seasonal"}

\item{life_history_params}{named vector (or list) with the numeric elements
"R0", "TR" (time to recovery) and "LP" (latent period)}

\item{vax_params}{named vector (or list) with the numeric elements "efficacy"
and "propn_vax0" (initial proportion of vaccinated individuals; assumed constant
across location, age and risk groups)}

\item{seasonality_params}{list of seasonality parameters.
contains the elements tdelay (0 <= tdelay <= 364): shifts the seasonality function - changing this effectively changes the seed time.
tdelay = 0 is seed at t = 0 in sinusoidal curve, roughly start of autumn in Northern hemisphere
days_per_block: average seasonality over blocks of this many days
amp: amplitude of seasonality}

\item{time_params}{list of parameters to do with time steps in simulation.}

\item{vax_alloc_period}{allocate vaccines once every this many timesteps}

\item{processed_inputs}{list. output of setup_inputs function}

\item{n_runs}{number of repeats to run}

\item{calculate_summaries_func}{string specifying what function to use to calculate summaries.}

\item{other_info}{list which provides information needed to calculate the summaries}
}
\value{
a list of lists containing results of main simulation loop (see main_simulation). The length of this list is equal to n_runs
}
\description{
function to set up simulation from inputs, then run
}

