% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_tools.R
\name{setup_cluster}
\alias{setup_cluster}
\title{Cluster setup}
\usage{
setup_cluster(user, expire = 1e-10)
}
\arguments{
\item{user}{a user identifier so that each user can specify their cluster options.}

\item{expire}{pass to the `provisionr::package_sources` function, deciding how long must elapse before the contexts folder is forced to update}
}
\value{
a didehpc::queue_didehpc object
}
\description{
This function sets up an object linking to the DIDE cluster in a very crude way. Each user will need to implement their own version of this that returns a valid didehpc object to submit jobs to.
}
