% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_inputs}
\alias{setup_inputs}
\title{process inputs to run simulation}
\usage{
setup_inputs(simulation_flags, life_history_params, travel_params, seed_params,
  user_specified_cum_vax_pool_func, vax_production_params,
  user_specified_vax_alloc_func, vax_allocation_params)
}
\arguments{
\item{simulation_flags}{named vector (or list) with the logical elements "normaliseTravel"
"seasonal" and "real_data"}

\item{life_history_params}{named vector (or list) with the numeric elements
"R0", "TR" (time to recovery) and "LP" (latent period)}

\item{travel_params}{list of parameters relating to travel. This should have the following elements: 1) epsilon, which scales the off-diagonals of the travel matrix}

\item{seed_params}{named vector (or list) of parameters to do with seeding the pandemic.
contains the elements seedCountries: vector of country names in which to seed}

\item{user_specified_cum_vax_pool_func}{function or character string of function to produce vaccine pool}

\item{vax_production_params}{named vector (or list) with named elements matching the arguments of user_specified_cum_vax_pool_func}

\item{user_specified_vax_alloc_func}{function or character string of function to allocate vaccines}

\item{vax_allocation_params}{named vector (or list) with named elements matching the arguments of user_specified_vax_alloc_func}

\item{time_params}{list of parameters to do with time steps in simulation.}

\item{seasonality_params}{list of seasonality parameters.
contains the elements tdelay (0 <= tdelay <= 364): shifts the seasonality function - changing this effectively changes the seed time.
tdelay = 0 is seed at t = 0 in sinusoidal curve, roughly start of autumn in Northern hemisphere
division: Average seasonality into this many blocks of time
amp: amplitude of seasonality}

\item{vax_params}{named vector (or list) with the numeric elements "efficacy"
and "propn_vax0" (initial proportion of vaccinated individuals; assumed constant
across location, age and risk groups)}
}
\value{
a list of arguments to go into run_simulation
}
\description{
process inputs to run simulation
}

