% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_populations}
\alias{setup_populations}
\title{function to setup synthetic population sizes for each location, age, risk group, 
and labels associated with these}
\usage{
setup_populations(popn_size, n_countries, age_propns, risk_propns, risk_factors)
}
\arguments{
\item{popn_size}{numeric vector of length 1: population size of a country.
assumed to be same across countries.}

\item{n_countries}{numeric vector of length 1: number of countries}

\item{age_propns}{numeric vector: proportion of individuals in each age group.
assumed to be the same across countries.}

\item{risk_propns}{numeric vector: proportion of individuals in each age group.
assumed to be the same across countries and age groups.}

\item{risk_factors}{numeric vector: degree of increased susceptibility 
in each risk group.
assumed to be the same across countries and age groups.}
}
\value{
list with the following elements:
X: numeric vector of length n_groups = n_countries * n_ages * n_riskgroups
containing the population sizes in each country, age, risk group
labels: data frame containing three columns: the location, age, risk group
corresponding to each element of X.
nrow(labels) = length(X)
pop_size: numeric vector of length n_countries containing the total population
size in each country.
}
\description{
function to setup synthetic population sizes for each location, age, risk group, 
and labels associated with these
}
