% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_populations_real_data}
\alias{setup_populations_real_data}
\title{function to setup population sizes for each location, age, risk group, 
and labels associated with these, from data}
\usage{
setup_populations_real_data(demography_filename, risk_propns, risk_factors,
  n_riskgroups)
}
\arguments{
\item{demography_filename}{character vector of length 1: file where demographic
data is located}

\item{risk_propns}{numeric vector: proportion of individuals in each age group.
assumed to be the same across countries and age groups.}

\item{risk_factors}{numeric vector: degree of increased susceptibility 
in each risk group.
assumed to be the same across countries and age groups.}
}
\value{
list with the following elements:
X: numeric vector of length n_groups = n_countries * n_ages * n_riskgroups
containing the population sizes in each country, age, risk group
labels: data frame containing three columns: the location, age, risk group
corresponding to each element of X.
nrow(labels) = length(X)
pop_size: numeric vector of length n_countries containing the total population
size in each country.
}
\description{
function to setup population sizes for each location, age, risk group, 
and labels associated with these, from data
}

