% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_travel_real_data}
\alias{setup_travel_real_data}
\title{function to construct travel matrix from data}
\usage{
setup_travel_real_data(travel_filename, pop_size, travel_params)
}
\arguments{
\item{travel_filename}{character vector of length 1: file where travel
data is located}

\item{pop_size}{numeric vector of length n_countries: population size in each
country}

\item{travel_params}{named vector/list containing the element "epsilon":
a numeric vector of length 1 which scales the off-diagonal terms of the 
travel matrix.  Roughly, the ratio of off-diagonal to on-diagonal term size.}
}
\value{
square travel matrix of side length n_countries (unnormalised)
}
\description{
function to construct travel matrix from data
}

