% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shorten_runs}
\alias{shorten_runs}
\title{make argument list with shorter number of runs}
\usage{
shorten_runs(args_list, n_runs_test)
}
\arguments{
\item{args_list}{an argument list to run simulations on or off the cluster}

\item{n_runs_test}{shorter number of runs}
}
\value{
an argument list to run simulations on or off the cluster
}
\description{
make argument list with shorter number of runs
}
