% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stochastic_simulation}
\alias{stochastic_simulation}
\title{Simulation}
\usage{
stochastic_simulation(S, I, R, beta, gamma, Kdelta, K1, Mm1, div, ndays,
  n_age_groups, n_countries)
}
\description{
Wow
}

