% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sum_age_risk_closure}
\alias{sum_age_risk_closure}
\title{returns a function to sum a state vector (e.g. the current number of susceptibles in each
country, age and risk group) across age and risk groups}
\usage{
sum_age_risk_closure(labels)
}
\arguments{
\item{labels}{data frame with n rows, and columns "Location", "RiskGroup" and "Age"}
}
\value{
a function to sum a state vector (e.g. the current number of susceptibles in each
country, age and risk group) across age and risk groups
}
\description{
returns a function to sum a state vector (e.g. the current number of susceptibles in each
country, age and risk group) across age and risk groups
}

