% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sum_list}
\alias{sum_list}
\title{sum over a list}
\usage{
sum_list(xs)
}
\arguments{
\item{xs}{list of objects that can be summed together}
}
\value{
the sum over the list
}
\description{
example use case: element-by-element sum over matrices in a list
}
