% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{thin_time_series}
\alias{thin_time_series}
\title{thin a time series matrix}
\usage{
thin_time_series(time_series_matrix, thin_every = 0, thin_integer = FALSE,
  thin_by_sum = FALSE)
}
\arguments{
\item{time_series_matrix}{matrix where each column corresponds to a given
time, and the column names are the times}

\item{thin_every}{an integer.  If 0, ignore.  Otherwise, thin the matrix
every this many columns}

\item{thin_integer}{a logical. If TRUE, thin the matrix to integer values of 
times.  If FALSE, ignore}

\item{thin_by_sum}{logical.  IF TRUE, thin by adding together columns (e.g.
if thin_every = 2, add togeher columns 1-2, columns 3-4 etc.)  This is useful
for thinning quantities such as incidence.  If FALSE, ignore.}
}
\value{
a thinned matrix with the same number of rows as time_series_matrix,
and fewer columns
}
\description{
thin a time series matrix
}
