% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{user_specified_cum_vax_pool_func}
\alias{user_specified_cum_vax_pool_func}
\title{example function of vaccine production:}
\usage{
user_specified_cum_vax_pool_func(vax_production_params, t)
}
\arguments{
\item{vax_production_params}{list of parameters for vaccine production}

\item{t}{scalar time}
}
\value{
needs to return a scalar: the number of vaccines ever produced up to time t
}
\description{
no vaccine produced until time vax_production_params[["detection_delay"]] + vax_production_params[["production_delay"]], then constant production rate until max number of doses ever made reached, then no production
}

