% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vax_alloc_funcs.R
\name{vaccinate_by_current_seasonal_alloc}
\alias{vaccinate_by_current_seasonal_alloc}
\title{allocate vaccines according to current seasonal allocation}
\usage{
vaccinate_by_current_seasonal_alloc(sum_age_risk_func, travel_matrix,
  vax_allocation_params, S, E, I, R, vax_pool)
}
\arguments{
\item{sum_age_risk_func}{a function which sums a vector across age and risk groups (created automatically in the code)}

\item{travel_matrix}{a square matrix with side length n_countries. Travel_matrix[x,y] is the proportion of time an individual in location x spends in location y}

\item{vax_allocation_params}{list of parameters for vaccine allocation}

\item{S}{numeric vector of length n. number of unvaccinated susceptibles.}

\item{E}{numeric vector of length n. number of unvaccinated exposed}

\item{I}{numeric vector of length n. number of unvaccinated infectious}

\item{R}{numeric vector of length n. number of unvaccinated recovered}

\item{vax_pool}{numeric vector of length 1. number of vaccines available}

\item{n}{is number of locations * number of age groups * number of risk groups}
}
\value{
n_vax_allocated numeric vector of length n_countries.  number of vaccines allocated to each country
}
\description{
allocate vaccines according to current seasonal allocation in each location, then
allocate uniformly within each country (without discriminating between
ages/risk groups/infection statuses)
}

