% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vcapply}
\alias{vcapply}
\title{vapply for characters}
\usage{
vcapply(X, FUN, ...)
}
\arguments{
\item{X}{X in vapply}

\item{FUN}{FUN in vapply}
}
\value{
vapply(X, FUN, character(1), ...)
}
\description{
vapply for characters
}
