% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{viapply}
\alias{viapply}
\title{vapply for integers}
\usage{
viapply(X, FUN, ...)
}
\arguments{
\item{X}{X in vapply}

\item{FUN}{FUN in vapply}
}
\value{
vapply(X, FUN, integer(1), ...)
}
\description{
vapply for integers
}
