% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vlapply}
\alias{vlapply}
\title{vapply for logicals}
\usage{
vlapply(X, FUN, ...)
}
\arguments{
\item{X}{X in vapply}

\item{FUN}{FUN in vapply}
}
\value{
vapply(X, FUN, logical(1), ...)
}
\description{
vapply for logicals
}
