% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vnapply}
\alias{vnapply}
\title{vapply for numerics}
\usage{
vnapply(X, FUN, ...)
}
\arguments{
\item{X}{X in vapply}

\item{FUN}{FUN in vapply}
}
\value{
vapply(X, FUN, numeric(1), ...)
}
\description{
vapply for numerics
}
