% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_funcs.R
\name{calibrating_amp_and_travel}
\alias{calibrating_amp_and_travel}
\title{submit job for amp and epsilon calibration}
\usage{
calibrating_amp_and_travel(runName, amp, epsilon, n_runs, time_params,
  seasonality_params, life_history_params, travel_params, simulation_flags,
  vax_params, vax_production_params, vax_allocation_params,
  user_specified_cum_vax_pool_func, user_specified_vax_alloc_func, seed_params,
  calculate_summaries_func, postprocessing_func, other_info, output_prefix)
}
\arguments{
\item{runName}{the base of the filename to save outputs to. Will be appended to "outputs/calibration_". DO NOT SPECIFY A FILE EXTENSION - THIS IS HANDLED IN THE FUNCTION}

\item{amp}{the amplitude of seasonal forcing}

\item{epsilon}{connectivity of off-diagonal elements in the travel matrix}

\item{n_runs}{number of simulations to run for this set of parameters}

\item{time_params}{list of parameters to do with time steps in simulation.}

\item{seasonality_params}{list of seasonality parameters.
contains the elements tdelay (0 <= tdelay <= 364): shifts the seasonality function - changing this effectively changes the seed time.
tdelay = 0 is seed at t = 0 in sinusoidal curve, roughly start of autumn in Northern hemisphere
division: Average seasonality into this many blocks of time
amp: amplitude of seasonality}

\item{life_history_params}{named vector (or list) with the numeric elements
"R0", "TR" (time to recovery) and "LP" (latent period)}

\item{travel_params}{list of parameters relating to travel. This should have the following elements: 1) epsilon, which scales the off-diagonals of the travel matrix}

\item{simulation_flags}{named vector (or list) with the logical elements "normaliseTravel"
and "seasonal"}

\item{vax_params}{named vector (or list) with the numeric elements "efficacy"
and "propn_vax0" (initial proportion of vaccinated individuals; assumed constant
across location, age and risk groups)}

\item{vax_production_params}{named vector (or list) with named elements matching the arguments of user_specified_cum_vax_pool_func}

\item{vax_allocation_params}{named vector (or list) with named elements matching the arguments of user_specified_vax_alloc_func}

\item{user_specified_cum_vax_pool_func}{function or character string of function to produce vaccine pool}

\item{user_specified_vax_alloc_func}{function or character string of function to allocate vaccines}

\item{seed_params}{named vector (or list) of parameters to do with seeding the pandemic.
contains the elements seedCountries: vector of country names in which to seed}

\item{calculate_summaries_func}{string specifying what function to use to calculate summaries.}

\item{postprocessing_func}{character string specifying function to do postprocessing}

\item{other_info}{list which provides any other information needed, such as to calculate the summaries}

\item{output_prefix}{character vector of length 1.  Prefix for output filenames}
}
\value{
returns TRUE if the routine runs correctly. Will create a .csv file of peak time summaries for this run, and a .png plotting the distribution of peak times by country. The output filenames are:
1) outputs/calibration_***runName***_data.csv
2) outputs/calibration_***runName***_plot.png
}
\description{
Function used to submit jobs, solving the global model with specified parameters. This particular function is intended to test different values of amp and epsilon, which is why they are amongst the first arguments. To write your own function which can be submitted to the cluster, simply place the parameters that you want to change as the first few arguments.
}
