% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_to_run_on_cluster.R
\name{postprocessing_simple_save}
\alias{postprocessing_simple_save}
\title{just save summary statistics from simulations as an .rds file}
\usage{
postprocessing_simple_save(res_list, runName)
}
\arguments{
\item{res_list}{summary statistics for each run and processed inputs produced by run_fixed_params}

\item{runName}{character string to make filename out of}
}
\description{
just save summary statistics from simulations as an .rds file
}
