% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{setup_cluster}
\alias{setup_cluster}
\title{Cluster setup}
\usage{
setup_cluster(user)
}
\arguments{
\item{user}{a user identifier so that each user can specify their cluster options.}
}
\value{
a didehpc::queue_didehpc object
}
\description{
This function sets up an object linking to the DIDE cluster in a very crude way. Each user will need to implement their own version of this that returns a valid didehpc object to submit jobs to.
}
